Loading typeclasses before everything else
